refactor: move ejb, interface and rest for each entity to respective folders under 'services'
