--- src/java/crypto/package-info.java
@@ -1,4 +1,0 @@
-/*
- * This package to contains classes for password ecryption and hashing.
- */
-package crypto;
