+++ src/java/services/animal/package-info.java
@@ -1,0 +1,6 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package services.animal;
