+++ src/java/services/user/IUserEjb.java
@@ -1,0 +1,24 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package services.user;
+
+import entities.UserEntity;
+import exceptions.CreateException;
+import exceptions.ReadException;
+import javax.ejb.Local;
+
+/**
+ *
+ * @author Ander
+ */
+@Local
+public interface IUserEjb {
+
+//    public void setUser(UserEntity user) throws CreateException;
+//
+//    public UserEntity getUser(String email) throws ReadException;
+    
+}
